+++ hypnotoad/cases/tokamak.py
@@ -53,6 +53,33 @@ class TokamakEquilibrium(Equilibrium):
             ),
             value_type=bool,
         ),
+        leg_refine_atol=WithMeta(
+            1.0e-5,
+            doc=(
+                "Tolerance used when iteratively refining the position of start points "
+                "of legs around an X-point."
+            ),
+            value_type=float,
+            check_all=is_positive,
+        ),
+        leg_refine_maxits=WithMeta(
+            1000,
+            doc=(
+                "Maximum number of iterations for iterative refinement of start points "
+                "of legs around an X-point."
+            ),
+            value_type=int,
+            check_all=is_positive,
+        ),
+        leg_trace_atol=WithMeta(
+            lambda options: 0.01 * options.refine_atol,
+            doc=(
+                "Tolerance used when calling scipy.integrate.solve_ivp to follow the "
+                "divertor legs. Default is 0.01*refine_atol."
+            ),
+            value_type=float,
+            check_all=is_positive,
+        ),
         nx_core=WithMeta(
             5,
             doc="Number of radial points in the core",
@@ -257,6 +284,24 @@ class TokamakEquilibrium(Equilibrium):
             value_type=float,
             check_all=[is_positive, lambda x: x < 1.0],
         ),
+        xpoint_refine_atol=WithMeta(
+            1.0e-6,
+            doc=(
+                "Tolerance used when iteratively refining the position of an X-point "
+                "that has been identified."
+            ),
+            value_type=float,
+            check_all=is_positive,
+        ),
+        xpoint_refine_maxits=WithMeta(
+            1000,
+            doc=(
+                "Maximum number of iterations for iterative refinement of the position "
+                "of an X-point that has been identified."
+            ),
+            value_type=int,
+            check_all=is_positive,
+        ),
     )
 
     def __init__(
@@ -396,7 +441,13 @@ class TokamakEquilibrium(Equilibrium):
 
         # Find critical points (O- and X-points)
         R2D, Z2D = np.meshgrid(R1D, Z1D, indexing="ij")
-        opoints, xpoints = critical.find_critical(R2D, Z2D, psi2D)
+        opoints, xpoints = critical.find_critical(
+            R2D,
+            Z2D,
+            psi2D,
+            self.user_options.xpoint_refine_atol,
+            self.user_options.xpoint_refine_maxits,
+        )
         if len(opoints) == 0:
             warnings.warn("No O-points found in TokamakEquilibrium input")
         else:
@@ -471,7 +522,8 @@ class TokamakEquilibrium(Equilibrium):
         inds = np.nonzero((psivals[1:] - psi_sep) * (psivals[:-1] - psi_sep) < 0.0)[0]
 
         # Currently only handle standard X-points (no snowflakes)
-        assert len(inds) == 2
+        if len(inds) != 2:
+            raise ValueError("Found more than 2 divertor legs")
 
         # Divide-and-conquer to get a points on the leg
         # This goes into a list leg_points = [(r,z),..]
@@ -489,7 +541,8 @@ class TokamakEquilibrium(Equilibrium):
             s2 = 1.0
             psi1 = psivals[ind]  # s = 0
 
-            while s2 - s1 > 1e-5:
+            count = 0
+            while s2 - s1 > self.user_options.leg_refine_atol:
                 smid = 0.5 * (s1 + s2)
                 psi_mid = self.psi(r0 + smid * dr, z0 + smid * dz)
 
@@ -499,6 +552,14 @@ class TokamakEquilibrium(Equilibrium):
                 else:
                     psi1 = psi_mid
                     s1 = smid
+                count += 1
+                if count > self.user_options.leg_refine_maxits:
+                    raise ValueError(
+                        f"Failed to find leg start position. Error {s2 - s1} is "
+                        f"greater than tolerance "
+                        f"{self.user_options.leg_refine_atol} after "
+                        f"{self.user_options.leg_refine_maxits} iterations."
+                    )
             smid = 0.5 * (s1 + s2)
             r = r0 + smid * dr
             z = z0 + smid * dz
@@ -529,7 +590,13 @@ class TokamakEquilibrium(Equilibrium):
             pos = leg  # Starting position
             while True:
                 # Integrate a distance "step" along the leg
-                solve_result = solve_ivp(dpos_dl, (0.0, step), pos)
+                solve_result = solve_ivp(
+                    dpos_dl,
+                    (0.0, step),
+                    pos,
+                    rtol=0.0,
+                    atol=self.user_options.leg_trace_atol,
+                )
                 newpos = (solve_result.y[0][1], solve_result.y[1][1])
 
                 # Check if we have crossed the boundary
@@ -587,7 +654,8 @@ class TokamakEquilibrium(Equilibrium):
         - self.regions         OrderedDict of EquilibriumRegion objects
 
         """
-        assert self.psi_axis is not None
+        if self.psi_axis is None:
+            raise ValueError("psi_axis has not been set")
 
         # psi values
         def psinorm_to_psi(psinorm):
@@ -635,7 +703,8 @@ class TokamakEquilibrium(Equilibrium):
         )
 
         # Check that there are only one or two left
-        assert 0 < len(self.x_points) <= 2
+        if not (0 < len(self.x_points) <= 2):
+            raise ValueError(f"Found unsupported number of X-points: {self.x_points}")
 
         if len(self.x_points) == 1:
             # Generate the specifications for a lower or upper single null
@@ -670,7 +739,11 @@ class TokamakEquilibrium(Equilibrium):
                         psi_vals    1D array of poloidal flux values. Length 2*nx+1
         connections    List of connections between regions
         """
-        assert len(self.x_points) == 1
+        if len(self.x_points) != 1:
+            raise ValueError(
+                f"Single-null case must have 1 X-point, got {self.x_points}"
+            )
+
         # Single null. Could be lower or upper
 
         # Find lines along the legs from X-point to target
@@ -846,7 +919,10 @@ class TokamakEquilibrium(Equilibrium):
         connections    List of connections between regions
         """
 
-        assert len(self.x_points) == 2
+        if len(self.x_points) != 2:
+            raise ValueError(
+                f"Double-null case must have 2 X-points, got {self.x_points}"
+            )
 
         if self.x_points[0].Z < self.o_point.Z:
             # Lower double null
@@ -860,8 +936,16 @@ class TokamakEquilibrium(Equilibrium):
         lower_psi = self.psi_sep[lower_xpt_ind]
         upper_psi = self.psi_sep[upper_xpt_ind]
 
-        assert np.isclose(lower_psi, self.psi(*lower_x_point))
-        assert np.isclose(upper_psi, self.psi(*upper_x_point))
+        if not np.isclose(lower_psi, self.psi(*lower_x_point)):
+            raise ValueError(
+                f"psi-value for lower separatrix {lower_psi} does not match value at "
+                "lower X-point {self.psi(*lower_x_point)}"
+            )
+        if not np.isclose(upper_psi, self.psi(*upper_x_point)):
+            raise ValueError(
+                f"psi-value for upper separatrix {upper_psi} does not match value at "
+                "upper X-point {self.psi(*upper_x_point)}"
+            )
 
         # Find lines along the legs from X-point to target
         lower_legs = self.findLegs(lower_x_point)
@@ -1436,7 +1520,8 @@ class TokamakEquilibrium(Equilibrium):
                     # Determine if poloidal flux is increasing or decreasing with radius
                     sign = np.sign(self.psi_sep[0] - self.psi_axis)
 
-                    assert region["psi"] is not None
+                    if region["psi"] is None:
+                        raise ValueError("No psi values in region")
                     leg_psi = region["psi"]
                     eqreg.pressure = lambda psi: self.pressure(
                         leg_psi + sign * abs(psi - leg_psi)
@@ -1495,10 +1580,11 @@ class TokamakEquilibrium(Equilibrium):
         def handler(self, *args):
             if isinstance(args[0], MultiLocationArray):
                 for arg in args[1:]:
-                    assert isinstance(arg, MultiLocationArray), (
-                        "if first arg is a MultiLocationArray, then others must be as "
-                        "well"
-                    )
+                    if not isinstance(arg, MultiLocationArray):
+                        raise ValueError(
+                            "if first arg is a MultiLocationArray, then others must be "
+                            "as well"
+                        )
                 result = MultiLocationArray(args[0].nx, args[0].ny)
 
                 if all(arg.centre is not None for arg in args):