+++ hypnotoad/utils/dct_interpolation.py
@@ -45,18 +45,20 @@ class DCT_2D:
         self.nZ = len(self.Zarray)
 
         # Assume constant spacing in R and Z
-        assert all(
+        if not all(
             numpy.abs(
                 self.Rarray - numpy.linspace(self.Rarray[0], self.Rarray[-1], self.nR)
             )
             < 1.0e-13
-        ), "grid spacing should be constant"
-        assert all(
+        ):
+            raise ValueError("grid spacing should be constant")
+        if not all(
             numpy.abs(
                 self.Zarray - numpy.linspace(self.Zarray[0], self.Zarray[-1], self.nZ)
             )
             < 1.0e-13
-        ), "grid spacing should be constant"
+        ):
+            raise ValueError("grid spacing should be constant")
 
         self.dR = self.Rarray[1] - self.Rarray[0]
         self.dZ = self.Zarray[1] - self.Zarray[0]
@@ -67,12 +69,12 @@ class DCT_2D:
         self.Zsize = self.Zarray[-1] - self.Zarray[0]
 
         # Check array sizes are compatible
-        assert (
-            self.nR == psiRZ.shape[1]
-        ), "size of R-direction should match size of columns of psiRZ"
-        assert (
-            self.nZ == psiRZ.shape[0]
-        ), "size of Z-direction should match size of rows of psiRZ"
+        if self.nR != psiRZ.shape[1]:
+            raise ValueError(
+                "size of R-direction should match size of columns of psiRZ"
+            )
+        if self.nZ != psiRZ.shape[0]:
+            raise ValueError("size of Z-direction should match size of rows of psiRZ")
 
         self.psiDCT = dct(dct(psiRZ, axis=0), axis=1)
 
@@ -89,17 +91,15 @@ class DCT_2D:
 
     def __call__(self, R, Z):
         if not isinstance(R, MultiLocationArray):
-            assert not isinstance(
-                Z, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if isinstance(Z, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             R = numpy.array(R)
             Z = numpy.array(Z)
 
             # check inputs are compatible
-            assert len(R.shape) == len(
-                Z.shape
-            ), "input R and Z should have same number of dimensions"
+            if len(R.shape) != len(Z.shape):
+                raise ValueError("input R and Z should have same number of dimensions")
 
         # calculate values in index space
         iR = (R - self.Rmin) / self.Rsize * (self.nR - 1)
@@ -120,9 +120,8 @@ class DCT_2D:
                 return it.operands[2]
 
         if isinstance(iR, MultiLocationArray):
-            assert isinstance(
-                iZ, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if not isinstance(iZ, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             result = MultiLocationArray(R.nx, R.ny)
             if iR.centre is not None and iZ.centre is not None:
@@ -143,17 +142,15 @@ class DCT_2D:
 
     def ddR(self, R, Z):
         if not isinstance(R, MultiLocationArray):
-            assert not isinstance(
-                Z, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if isinstance(Z, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             R = numpy.array(R)
             Z = numpy.array(Z)
 
             # check inputs are compatible
-            assert len(R.shape) == len(
-                Z.shape
-            ), "input R and Z should have same number of dimensions"
+            if len(R.shape) != len(Z.shape):
+                raise ValueError("input R and Z should have same number of dimensions")
 
         # calculate values in index space
         iR = (R - self.Rmin) / self.Rsize * (self.nR - 1)
@@ -176,9 +173,8 @@ class DCT_2D:
                 return it.operands[2]
 
         if isinstance(iR, MultiLocationArray):
-            assert isinstance(
-                iZ, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if not isinstance(iZ, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             result = MultiLocationArray(R.nx, R.ny)
             if iR.centre is not None and iZ.centre is not None:
@@ -199,17 +195,15 @@ class DCT_2D:
 
     def ddZ(self, R, Z):
         if not isinstance(R, MultiLocationArray):
-            assert not isinstance(
-                Z, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if isinstance(Z, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             R = numpy.array(R)
             Z = numpy.array(Z)
 
             # check inputs are compatible
-            assert len(R.shape) == len(
-                Z.shape
-            ), "input R and Z should have same number of dimensions"
+            if len(R.shape) != len(Z.shape):
+                raise ValueError("input R and Z should have same number of dimensions")
 
         # calculate values in index space
         iR = (R - self.Rmin) / self.Rsize * (self.nR - 1)
@@ -232,9 +226,8 @@ class DCT_2D:
                 return it.operands[2]
 
         if isinstance(iR, MultiLocationArray):
-            assert isinstance(
-                iZ, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if not isinstance(iZ, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             result = MultiLocationArray(R.nx, R.ny)
             if iR.centre is not None and iZ.centre is not None:
@@ -255,17 +248,15 @@ class DCT_2D:
 
     def d2dR2(self, R, Z):
         if not isinstance(R, MultiLocationArray):
-            assert not isinstance(
-                Z, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if isinstance(Z, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             R = numpy.array(R)
             Z = numpy.array(Z)
 
             # check inputs are compatible
-            assert len(R.shape) == len(
-                Z.shape
-            ), "input R and Z should have same number of dimensions"
+            if len(R.shape) != len(Z.shape):
+                raise ValueError("input R and Z should have same number of dimensions")
 
         # calculate values in index space
         iR = (R - self.Rmin) / self.Rsize * (self.nR - 1)
@@ -287,9 +278,8 @@ class DCT_2D:
                 return it.operands[2]
 
         if isinstance(iR, MultiLocationArray):
-            assert isinstance(
-                iZ, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if not isinstance(iZ, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             result = MultiLocationArray(R.nx, R.ny)
             if iR.centre is not None and iZ.centre is not None:
@@ -310,17 +300,15 @@ class DCT_2D:
 
     def d2dZ2(self, R, Z):
         if not isinstance(R, MultiLocationArray):
-            assert not isinstance(
-                Z, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if isinstance(Z, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             R = numpy.array(R)
             Z = numpy.array(Z)
 
             # check inputs are compatible
-            assert len(R.shape) == len(
-                Z.shape
-            ), "input R and Z should have same number of dimensions"
+            if len(R.shape) != len(Z.shape):
+                raise ValueError("input R and Z should have same number of dimensions")
 
         # calculate values in index space
         iR = (R - self.Rmin) / self.Rsize * (self.nR - 1)
@@ -342,9 +330,8 @@ class DCT_2D:
                 return it.operands[2]
 
         if isinstance(iR, MultiLocationArray):
-            assert isinstance(
-                iZ, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if not isinstance(iZ, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             result = MultiLocationArray(R.nx, R.ny)
             if iR.centre is not None and iZ.centre is not None:
@@ -365,17 +352,15 @@ class DCT_2D:
 
     def d2dRdZ(self, R, Z):
         if not isinstance(R, MultiLocationArray):
-            assert not isinstance(
-                Z, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if isinstance(Z, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             R = numpy.array(R)
             Z = numpy.array(Z)
 
             # check inputs are compatible
-            assert len(R.shape) == len(
-                Z.shape
-            ), "input R and Z should have same number of dimensions"
+            if len(R.shape) != len(Z.shape):
+                raise ValueError("input R and Z should have same number of dimensions")
 
         # calculate values in index space
         iR = (R - self.Rmin) / self.Rsize * (self.nR - 1)
@@ -400,9 +385,8 @@ class DCT_2D:
                 return it.operands[2]
 
         if isinstance(iR, MultiLocationArray):
-            assert isinstance(
-                iZ, MultiLocationArray
-            ), "if R is a MultiLocationArray, then Z must be as well"
+            if not isinstance(iZ, MultiLocationArray):
+                raise ValueError("if R is a MultiLocationArray, then Z must be as well")
 
             result = MultiLocationArray(R.nx, R.ny)
             if iR.centre is not None and iZ.centre is not None: